+++ examples/editor_demo/src/editor_ext.rs
@@ -82,8 +82,8 @@ fn setup_drag(ec: &mut EntityCommands, widget_id: Entity)
 //-------------------------------------------------------------------------------------------------------------------
 
 // TODO: how to make this generic? might need a trait
-fn make_draggable_field_widget<'a>(
-    l: &mut LoadedScene<'a, UiBuilder<'a, Entity>>,
+fn make_draggable_field_widget(
+    l: &mut LoadedSceneUi,
     name: &'static str,
     initial_value: f32,
     bounds: RangeInclusive<f32>,